+++ web/app/models/usage_snapshot.rb
@@ -1,0 +1,2 @@
+class UsageSnapshot < ApplicationRecord
+end